requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
jsonschema>=4.0.0
numpy>=1.20.0
asyncio>=3.4.3
# Optional but recommended for vector-based memory system
sentence-transformers>=2.2.0
